refactor: normalizeLink loose mode
