+++ packages/core/src/node/index.ts
@@ -1,6 +1,5 @@
 export { build } from './build';
 export { dev } from './dev';
 export { remarkFileCodeBlock } from './mdx/remarkPlugins/fileCodeBlock';
-
-export { remarkPluginNormalizeLink } from './mdx/remarkPlugins/normalizeLink';
+export { remarkNormalizeLink } from './mdx/remarkPlugins/normalizeLink';
 export { serve } from './serve';
