+++ packages/core/src/node/mdx/remarkPlugins/image.test.ts
@@ -5,11 +5,9 @@ describe('mdx', () => {
   it('basic', async () => {
     const result = await compile({
       source: `
-[link1](./test1.md)
+![alt1](./test3.jpg)
 
-{/* jsx link will not be transformed */}
-
-<a href="./test2">link2</a>
+<img src="./test4.png" alt="alt2" />
 `,
       checkDeadLinks: false,
       docDirectory: '/usr/rspress-project/docs',
